improve memory domain tests
Contexts are allowed to allocate a larger amount of memory than they were asked for. This specifically is a requirement for the Wasm backends, where the Wasm standard dictates the Wasm memory to be a multiple of the Wasm page size (64Kib), so we need to round up to the next multiple. This mainly adapts the OOB tests to take the actual context size into account.